+++ client/image_create.go
@@ -2,6 +2,7 @@ package client // import "github.com/docker/docker/client"
 
 import (
 	"context"
+	"encoding/json"
 	"io"
 	"net/http"
 	"net/url"
@@ -10,6 +11,7 @@ import (
 	"github.com/distribution/reference"
 	"github.com/docker/docker/api/types/image"
 	"github.com/docker/docker/api/types/registry"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
 
 // ImageCreate creates a new image based on the parent options.
@@ -37,4 +39,22 @@ func (cli *Client) tryImageCreate(ctx context.Context, query url.Values, registr
 	return cli.post(ctx, "/images/create", query, nil, http.Header{
 		registry.AuthHeader: {registryAuth},
 	})
+}
+
+func (cli *Client) ImageCreateFromOCIIndex(ctx context.Context, ref reference.NamedTagged, index ocispec.Index) (ocispec.Descriptor, error) {
+	query := url.Values{}
+	query.Set("fromJSON", "1")
+	query.Set("repo", ref.Name())
+	query.Set("tag", ref.Tag())
+
+	resp, err := cli.post(ctx, "/images/create", query, index, nil)
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+
+	var desc ocispec.Descriptor
+	if err := json.NewDecoder(resp.body).Decode(&desc); err != nil {
+		return ocispec.Descriptor{}, err
+	}
+	return desc, nil
 }